+++ R/data-combined.R
@@ -340,11 +340,11 @@ DataCombined <- R6::R6Class(
 
     #' @description
     #'
-    #' A method to extract a dataframe of simulated and/or observed data
+    #' A method to extract a tibble data frame of simulated and/or observed data
     #' (depending on instances of which classes have been added to the object).
     #'
     #' Note that the order in which you enter different object doesn't matter
-    #' because the returned dataframe is arranged alphabetically by dataset
+    #' because the returned data frame is arranged alphabetically by dataset
     #' name.
     #'
     #' @return